# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the Prim API",
        "title": "Prim API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/admin/brands": {
            "get": {
                "description": "list all brands in pages for administration",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Brand"
                ],
                "summary": "list all brands including soft-deleted ones (admin)",
                "parameters": [
                    {
                        "type": "integer",
                        "example": 1,
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "example": 10,
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "television",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "example": [
                            "name",
                            "-created_at"
                        ],
                        "name": "sort",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.PaginatedResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/brand.AdminBrandResponse"
                                            }
                                        },
                                        "meta": {
                                            "$ref": "#/definitions/api.Page"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "create a new product brand",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Brand"
                ],
                "summary": "create a new product brand",
                "parameters": [
                    {
                        "description": "brand data",
                        "name": "brand",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/brand.CreateBrandRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.DataResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/brand.BrandResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/brands/{id}": {
            "get": {
                "description": "get brand by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Brand"
                ],
                "summary": "get brand by id",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Brand ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.DataResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/brand.BrandResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "soft-delete an active brand by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Brand"
                ],
                "summary": "soft-delete a brand",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Brand ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "description": "update specific brand fields by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Brand"
                ],
                "summary": "update a brand",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Brand ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "brand update payload",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/brand.UpdateBrandRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/brands": {
            "get": {
                "description": "list public active brands in pages",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Brand"
                ],
                "summary": "list public active brands",
                "parameters": [
                    {
                        "type": "integer",
                        "example": 1,
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "example": 10,
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "television",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "example": [
                            "name",
                            "-created_at"
                        ],
                        "name": "sort",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.PaginatedResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/brand.BrandResponse"
                                            }
                                        },
                                        "meta": {
                                            "$ref": "#/definitions/api.Page"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.DataResponse": {
            "type": "object",
            "properties": {
                "data": {}
            }
        },
        "api.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "CODE_ERROR"
                },
                "message": {
                    "type": "string",
                    "example": "an error occured"
                }
            }
        },
...
        "api.MessageResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "the request sent successfully"
                }
            }
        },
        "api.Page": {
            "type": "object",
            "properties": {
                "has_next": {
                    "type": "boolean",
                    "example": true
                },
                "has_previous": {
                    "type": "boolean",
                    "example": false
                },
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "page_size": {
                    "type": "integer",
                    "example": 10
                },
                "total_items": {
                    "type": "integer",
                    "example": 20
                },
                "total_pages": {
                    "type": "integer",
                    "example": 2
                }
            }
        },
        "api.PaginatedResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "meta": {}
            }
        },
...
        "brand.AdminBrandResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2026-07-01T05:04:38Z"
                },
                "deleted_at": {
                    "type": "string",
                    "example": "2026-07-01T05:04:38Z"
                },
                "id": {
                    "type": "string",
                    "example": "358b2e03-0b3f-40a4-8163-ebed0cb252ee"
                },
                "link": {
                    "type": "string",
                    "example": "https://nvidia.com"
                },
                "logo_storage_object_id": {
                    "type": "string",
                    "example": "358b2e03-0b3f-40a4-8163-ebed0cb252ee"
                },
                "name": {
                    "type": "string",
                    "example": "nvidia"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2026-07-01T05:04:38Z"
                }
            }
        },
        "brand.BrandResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2026-07-01T05:04:38Z"
                },
                "id": {
                    "type": "string",
                    "example": "358b2e03-0b3f-40a4-8163-ebed0cb252ee"
                },
                "link": {
                    "type": "string",
                    "example": "https://nvidia.com"
                },
                "logo_storage_object_id": {
                    "type": "string",
                    "example": "358b2e03-0b3f-40a4-8163-ebed0cb252ee"
                },
                "name": {
                    "type": "string",
                    "example": "nvidia"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2026-07-01T05:04:38Z"
                }
            }
        },
        "brand.CreateBrandRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "link": {
                    "type": "string",
                    "example": "https://apple.com"
                },
                "name": {
                    "type": "string",
                    "example": "apple"
                }
            }
        },
        "brand.UpdateBrandRequest": {
            "type": "object",
            "properties": {
                "link": {
                    "type": "string",
                    "example": "https://apple.com"
                },
                "logo_storage_object_id": {
                    "type": "string",
                    "example": "358b2e03-0b3f-40a4-8163-ebed0cb252ee"
                },
                "name": {
                    "type": "string",
                    "example": "apple"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "security": [
        {
            "BearerAuth": []
        }
    ]
}